added extra code for easier testing without vp-api
