+++ config.js
@@ -18,6 +18,7 @@ if ([DOMAIN, VP_API_CLIENT_ID, VP_API_CLIENT_SECRET].some((envVar) => !envVar)) 
 
 const ENABLE_DEBUG_FILE_WRITING = isTruthy(process.env.ENABLE_DEBUG_FILE_WRITING);
 const ENABLE_SENDING_TO_VP_API = isTruthy(process.env.ENABLE_SENDING_TO_VP_API);
+const ENABLE_ALWAYS_CREATE_PARLIAMENT_FLOW = isTruthy(process.env.ENABLE_ALWAYS_CREATE_PARLIAMENT_FLOW);
 
 const DOCUMENT_TYPES = {
   BESLISSINGSFICHE: 'http://themis.vlaanderen.be/id/concept/document-type/e807feec-1958-46cf-a558-3379b5add49e',
@@ -58,4 +59,5 @@ export {
   PARLIAMENT_FLOW_STATUSES,
   ENABLE_DEBUG_FILE_WRITING,
   ENABLE_SENDING_TO_VP_API,
+  ENABLE_ALWAYS_CREATE_PARLIAMENT_FLOW
 };
