+++ lib/parliament-flow.js
@@ -279,6 +279,38 @@ WHERE {
   return pieces;
 }
 
+async function createOrUpdateParliamentFlow(responseJson, decisionmakingFlowUri, pieces, currentUser, comment, isComplete) {
+
+  const parliamentId = responseJson.pobj;
+  pieces.forEach((piece) => {
+    piece.files.forEach((file) => {
+      const parliamentId = responseJson.files.find((r) => r.id === file.uri)?.pfls;
+      if (parliamentId) {
+        file.parliamentId = parliamentId;
+      }
+    });
+  });
+
+  let { parliamentFlow, parliamentSubcase } =
+    await getParliamentFlowAndSubcase(decisionmakingFlowUri);
+
+  parliamentFlow ??= await createParliamentFlow(
+    parliamentId,
+    decisionmakingFlowUri
+  );
+  parliamentSubcase ??= await createParliamentSubcase(parliamentFlow);
+
+  const submissionActivity = await createSubmissionActivity(parliamentSubcase, currentUser, comment);
+  await createSubmittedPieces(submissionActivity, pieces)
+
+  await updateParliamentFlowStatus(
+    parliamentFlow,
+    isComplete
+      ? PARLIAMENT_FLOW_STATUSES.COMPLETE
+      : PARLIAMENT_FLOW_STATUSES.INCOMPLETE,
+  );
+}
+
 export {
   getParliamentFlowAndSubcase,
   createParliamentFlow,
@@ -287,4 +319,5 @@ export {
   createSubmissionActivity,
   createSubmittedPieces,
   enrichPiecesWithPreviousSubmissions,
+  createOrUpdateParliamentFlow
 };
